Scatter using canvas
