+++ scatter/style.scss
@@ -1,4 +1,40 @@
 .scatter-plot {
+    position: relative;
+}
+
+.scatter-plot.canvas {
+    position: absolute;
+    pointer-events: all;
+    z-index: 0;
+}
+
+.scatter-plot.svg {
+    position: absolute;
+    z-index: -1;
+    pointer-events: none;
+    -webkit-transform: translate3d(0,0,0);
+}
+
+.scatter-plot.rect {
+    fill: #ffffff;
+    z-index: 100;
+}
+
+.scatter-plot.brush-container {
+    z-index : 0;
+    pointer-events: none;
+}
+
+.brush {
+    pointer-events: none;
+}
+
+.brush .extent {
+    fill-opacity: .1;
+    shape-rendering: crispEdges;
+}
+
+.scatter-plot {
 
     .plot {
         fill: #ffffff;
